requirements.txt
torch
torchvision
flask
pillow
pandas
flask-cors
